move common code to util func
